Multiple networks from a single edgelist with a typed attribute

Create a list of igraph objects from an edgelist according to a type attribute

Usage

graph_from_multi_edgelist(
  d,
  from = NULL,
  to = NULL,
  type = NULL,
  weight = NULL,
  directed = FALSE
)

Arguments

d

data frame.

from

column name of sender. If NULL, defaults to first column.

to

column of receiver. If NULL, defaults to second column.

type

type attribute to split the edgelist. If NULL, defaults to third column.

weight

optional column name of edge weights. Ignored if NULL.

directed

logical scalar, whether or not to create a directed graph.

Value

list of igraph objects.

Author

David Schoch

Examples

library(igraph)
d <- data.frame(from=rep(c(1,2,3),3),to=rep(c(2,3,1),3),
                type=rep(c("a","b","c"),each=3),weight=1:9)
graph_from_multi_edgelist(d,"from","to","type","weight")
#> $a
#> IGRAPH 10c6651 UNW- 3 3 -- 
#> + attr: name (v/c), weight (e/n), type (e/c)
#> + edges from 10c6651 (vertex names):
#> [1] 1--2 2--3 1--3
#> 
#> $b
#> IGRAPH 4acab7d UNW- 3 3 -- 
#> + attr: name (v/c), weight (e/n), type (e/c)
#> + edges from 4acab7d (vertex names):
#> [1] 1--2 2--3 1--3
#> 
#> $c
#> IGRAPH b088746 UNW- 3 3 -- 
#> + attr: name (v/c), weight (e/n), type (e/c)
#> + edges from b088746 (vertex names):
#> [1] 1--2 2--3 1--3
#>